requirements.txt
absl-py==0.9.0
apt-xapian-index==0.49
apturl==0.5.2
asgiref==3.2.3
asn1crypto==0.24.0
astor==0.8.1
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.6
beautifulsoup4==4.6.0
bleach==3.1.1
blinker==1.4
bootstrap4==0.1.0
boto==2.49.0
boto3==1.12.16
botocore==1.15.16
branca==0.4.0
Brlapi==0.6.7
CacheControl==0.12.6
cachetools==4.0.0
cachy==0.3.0
cairocffi==0.9.0
certifi==2018.8.24
cffi==1.12.3
chardet==3.0.4
cleo==0.7.6
Click==7.0
clikit==0.4.1
colorama==0.4.3
command-not-found==0.3
configparser==4.0.2
crayons==0.3.0
cryptography==2.6.1
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.12
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distro==1.3.0
distro-info===0.21ubuntu4
dj-database-url==0.5.0
Django==3.0.3
django-allauth==0.41.0
django-bootstrap4==1.1.1
django-ckeditor==5.9.0
django-crispy-forms==1.8.1
django-debug-toolbar==2.2
django-ipware==2.1.0
django-js-asset==1.2.2
django-modelcluster==5.0.1
django-ranged-response==0.2.0
django-recaptcha==2.0.6
django-richtextfield==1.5.0
django-simple-captcha==0.5.12
django-taggit==1.2.0
django-tinymce==2.8.0
django-treebeard==4.3.1
django-user-agents==0.4.0
djangorestframework==3.11.0
docutils==0.15.2
draftjs-exporter==2.1.7
duplicity==0.8.4
entrypoints==0.3
f.lux-indicator-applet==1.1.11rc0
fasteners==0.12.0
Flask==1.0.2
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.1
folium==0.10.1
future==0.16.0
gast==0.2.2
gensim==3.8.1
geographiclib==1.50
geopy==1.21.0
google-auth==1.11.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gpg==1.12.0
grpcio==1.27.2
h5py==2.10.0
html5lib==1.0.1
httplib2==0.11.3
idna==2.6
importlib-metadata==1.1.3
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
json5==0.9.2
jsonschema==3.2.0
jupyter-client==6.0.0
jupyter-core==4.6.3
jupyterlab==2.0.1
jupyterlab-server==1.0.7
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==20.0.1
keyrings.alt==3.1.1
kiwisolver==1.1.0
l18n==2018.5
language-selector==0.1
launchpadlib==1.10.7
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.10.0
macaroonbakery==1.2.3
Mako==1.0.7
Markdown==3.2.1
MarkupSafe==1.1.0
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
msgpack==1.0.0
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.4
nltk==3.4.5
notebook==6.0.3
numpy==1.18.1
oauthlib==3.0.1
olefile==0.46
opencv-python==4.2.0.32
opt-einsum==3.2.0
pandas==1.0.0
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.6.2
pastel==0.1.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.1.0
pipenv==11.9.0
pkginfo==1.5.0.1
plotly==4.5.3
ply==3.11
poetry==1.0.3
powerline-status==2.7
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.11.3
psutil==5.5.1
psycopg2==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.6.1
PyGObject==3.34.0
PyJWT==1.7.0
pylev==1.3.0
pylint==2.4.4
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
PyQt5==5.12.3
pyRFC3339==1.1
pyrsistent==0.14.11
PySocks==1.6.8
python-apt==1.9.0+ubuntu1.3
python-dateutil==2.7.3
python-debian==0.1.36
python3-openid==3.1.0
pytz==2017.3
pyxdg==0.25
PyYAML==5.1.2
pyzmq==19.0.0
reportlab==3.5.23
requests==2.21.0
requests-oauthlib==1.3.0
requests-toolbelt==0.8.0
requests-unixsocket==0.1.5
retrying==1.3.3
rsa==4.0
s3transfer==0.3.3
scikit-learn==0.22.1
scipy==1.4.1
screen-resolution-extra==0.0.0
seaborn==0.10.0
SecretStorage==2.3.1
selenium==3.141.0
Send2Trash==1.5.0
shellingham==1.3.2
simplejson==3.16.0
sip==4.19.18
six==1.12.0
sklearn==0.0
smart-open==1.9.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
SQLAlchemy==1.3.13
sqlparse==0.3.0
system-service==0.3
systemd-python==234
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tomlkit==0.5.8
torbrowser-launcher==0.3.2
tornado==6.0.4
tqdm==4.43.0
traitlets==4.3.3
typed-ast==1.4.1
ua-parser==0.9.0
ubuntu-advantage-tools==19.5
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.1.1
urllib3==1.24.1
usb-creator==0.3.7
user-agents==2.1
virtualenv==15.1.0
virtualenv-clone==0.3.0
wadllib==1.3.3
wagtail==2.8
wcwidth==0.1.8
webdriver-manager==2.3.0
webencodings==0.5.1
Werkzeug==0.16.1
whitenoise==5.0.1
Willow==1.3
wordcloud==1.6.0
wrapt==1.11.2
xcffib==0.8.1
xkit==0.0.0
zipp==3.0.0
